% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.manag.sigma.R
\name{data.manag.sigma}
\alias{data.manag.sigma}
\title{Management of data for residual part}
\usage{
data.manag.sigma(formGroup, formFixed, formRandom, data.long1)
}
\arguments{
\item{formGroup}{A formula which indicates the group variable}

\item{formFixed}{A formula which indicates the fixed effects for the longitudinal submodel}

\item{formRandom}{A formula which indicates the random effects for the longitudinal submodel}

\item{data.long1}{A dataframe with the longitudinal data}
}
\description{
Management of data for residual part
}
