% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.manag.surv.R
\name{data.manag.surv}
\alias{data.manag.surv}
\title{Management of survival data}
\usage{
data.manag.surv(formGroup, formSurv, data.long1)
}
\arguments{
\item{formGroup}{A formula which indicates the group variable}

\item{formSurv}{A formula which indicates the survival submodel}

\item{data.long1}{A dataframe with the longitudinal data}
}
\description{
Management of survival data
}
