% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict.lsjm_classicCR.R,
%   R/predict.lsjm_classicIDM.R, R/predict.lsjm_classicSingle.R,
%   R/predict.lsjm_covDepCR.R, R/predict.lsjm_covDepIDM.R,
%   R/predict.lsjm_covDepSingle.R, R/predict.lsjm_interintraCR.R,
%   R/predict.lsjm_interintraIDM.R, R/predict.lsjm_interintraSingle.R,
%   R/predict.lsmm_classic.R, R/predict.lsmm_covDep.R,
%   R/predict.lsmm_interintra.R
\name{predict.lsjm}
\alias{predict.lsjm}
\alias{predict.lsjm_classicCR}
\alias{predict.lsjm_classicIDM}
\alias{predict.lsjm_classicSingle}
\alias{predict.lsjm_covDepCR}
\alias{predict.lsjm_covDepIDM}
\alias{predict.lsjm_covDepSingle}
\alias{predict.lsjm_interintraCR}
\alias{predict.lsjm_interintraIDM}
\alias{predict.lsjm_interintraSingle}
\alias{predict.lsmm_classic}
\alias{predict.lsmm_covDep}
\alias{predict.lsmm_interintra}
\title{predict: Prediction of some quantities for each subjects}
\usage{
\method{predict}{lsjm_classicCR}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_classicIDM}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_classicSingle}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_covDepCR}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_covDepIDM}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_covDepSingle}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_interintraCR}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_interintraIDM}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsjm_interintraSingle}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsmm_classic}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsmm_covDep}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)

\method{predict}{lsmm_interintra}(object, which = "RE", Objectranef = NULL, data.long = NULL, ...)
}
\arguments{
\item{object}{Either a \code{lsmm} object or a \code{lsjm} object}

\item{which}{A vector of characters to indicate which predictions are computed.
\code{"RE"} corresponds to the random effects, \code{"Y"} to the marker and \code{"Cum"} to the
cumulative risk function(s) (only in the case of a \code{lsjm} object)}

\item{Objectranef}{Optional \code{ranef} object containing the predicted random effects for each
individual. The Default is NULL and the predict functions should compute the random effects.}

\item{data.long}{A dataframe containing the longitudinal data for making predictions.}

\item{...}{Further arguments}
}
\value{
A table for each type of prediction (RE/Y/Cum)
\describe{
\item{\code{predictRE}}{A table with predicted random effects for each subjects.}
\item{\code{predictY}}{A table with predicted marker and residual variability for each subjects
and at each measurement time.}
\item{\code{predictCum_01}}{A table for predicted cumulative hazard function for transition 0-1.}
\item{\code{predictCum_02}}{A table for predicted cumulative hazard function for transition 0-2.}
\item{\code{predictCum_12}}{A table for predicted cumulative hazard function for transition 1-2.}
\item{\code{grid.time.Cum}}{A vector of times for which the cumulative hazard functions are
computed.}
}
}
\description{
This function computes different predicted quantities. For a \code{lsmm} object, it is possible
to compute, for each subject, their random effects, their value of the marker and the residual
variability (for each measurement time)).
For a \code{lsjm} object we could also compute cumulative hazard function for each risk
transition.
}
\examples{

data <- data.frame(
         ID = rep(1:5, each = 3),
         time = rep(1:3, 5),
         y = rnorm(15),
         event = rep(rbinom(5,1,0.5), each = 3),
         time_event = rep(runif(5), each = 3)
         )


m0 <- lsmm(
   formFixed = y ~ time,
   formRandom = ~ time,
   formGroup = ~ ID,
   formVar = "standard",
   timeVar = "time",
   data.long = data,
   S1 = 5,
   S2 = 5,
   nproc = 1
   )

pred.m0 <- predict(m0, which = c("RE","Y"), data.long = data)


\donttest{
library(dplyr)

data(threeC)
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$SBP <- threeC$SBP/10
threeC <- threeC
threeC <- dplyr::group_by(threeC, ID, num.visit)
threeC <- dplyr::mutate(threeC, SBPvisit = mean(SBP))
threeC_ex1 <- threeC[!duplicated(threeC[, c("ID", "num.visit")]),
                                  c("ID", "SBPvisit", "age.visit65", "sex","dem","death",
                                    "age.first65","age.last65","age.final65","age0_65")]

m1 <- lsmm(formFixed = SBPvisit ~ age.visit65,
               formRandom = ~ age.visit65,
               formGroup = ~ ID,
               timeVar = 'age.visit65',
               data.long = threeC_ex1,
               formVar = "cov-dependent",
               formFixedVar = ~ age.visit65+sex,
               formRandomVar = ~ age.visit65,
               correlated_re = FALSE,
               S1 = 500,
               S2 = 1000,
               nproc = 1)

l1 <- lsjm(m1,
           survival_type = 'IDM',
           formSurv_01=~1,
           formSurv_02=~sex,
           formSurv_12=~sex,
           sharedtype_01 = c("value", "variability"),
           sharedtype_02 = c("value", "slope", "variability"),
           sharedtype_12 = c("value", "variability"),
           hazardBase_01 = "Weibull",
           hazardBase_02 = "Weibull",
           hazardBase_12 = "Splines",
           delta1=~dem,
           delta2=~death,
           Time_T =~age.final65,
           Time_L =~age.last65,
           Time_R =~age.first65,
           Time_T0 =~age0_65,
           formSlopeFixed =~1,
           formSlopeRandom =~1,
           index_beta_slope = c(2),
           index_b_slope = c(2),
           nb.knots.splines = c(0,0,1),
           S1 = 1000,
           S2 = 2000,
           nproc = 10)

pred.m1 <- predict(m1, which = c("RE","Y"), data.long = threeC_ex1)

pred.l1 <- predict(l1, which = c("RE","Y","Cum"), data.long = threeC_ex1)


}




}
