% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict.lsjm_classicCR.R,
%   R/predict.lsjm_covDepCR.R, R/predict.lsjm_covDepIDM.R,
%   R/predict.lsjm_covDepSingle.R, R/predict.lsjm_interintraCR.R,
%   R/predict.lsjm_interintraIDM.R, R/predict.lsmm_classic.R,
%   R/predict.lsmm_covDep.R, R/predict.lsmm_interintra.R
\name{predict}
\alias{predict}
\alias{predict.lsjm_classicCR}
\alias{predict.lsjm_covDepCR}
\alias{predict.lsjm_covDepIDM}
\alias{predict.lsjm_covDepSingle}
\alias{predict.lsjm_interintraCR}
\alias{predict.lsjm_interintraIDM}
\alias{predict.lsmm_classic}
\alias{predict.lsmm_covDep}
\alias{predict.lsmm_interintra}
\title{predict : predict, for a lsjm or lsmm object, the random effects, the current value, the residual variability and for each individuals and the cumulative hazard function for eahc events (lsjm object only)}
\usage{
\method{predict}{lsjm_classicCR}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_covDepCR}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_covDepIDM}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_covDepSingle}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_interintraCR}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_interintraIDM}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsmm_classic}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsmm_covDep}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsmm_interintra}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)
}
\description{
predict : predict, for a lsjm or lsmm object, the random effects, the current value, the residual variability and for each individuals and the cumulative hazard function for eahc events (lsjm object only)
}
