% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R, R/ranef.lsjm_classicCR.R,
%   R/ranef.lsjm_classicIDM.R, R/ranef.lsjm_classicSingle.R,
%   R/ranef.lsjm_covDepCR.R, R/ranef.lsjm_covDepIDM.R,
%   R/ranef.lsjm_covDepSingle.R, R/ranef.lsjm_interintraCR.R,
%   R/ranef.lsjm_interintraIDM.R, R/ranef.lsjm_interintraSingle.R,
%   R/ranef.lsmm_classic.R, R/ranef.lsmm_covDep.R, R/ranef.lsmm_interintra.R
\name{ranef}
\alias{ranef}
\alias{ranef.lsjm_classicCR}
\alias{ranef.lsjm_classicIDM}
\alias{ranef.lsjm_classicSingle}
\alias{ranef.lsjm_covDepCR}
\alias{ranef.lsjm_covDepIDM}
\alias{ranef.lsjm_covDepSingle}
\alias{ranef.lsjm_interintraCR}
\alias{ranef.lsjm_interintraIDM}
\alias{ranef.lsjm_interintraSingle}
\alias{ranef.lsmm_classic}
\alias{ranef.lsmm_covDep}
\alias{ranef.lsmm_interintra}
\title{ranef: the predicted random effects for all subjects}
\usage{
\method{ranef}{lsjm_classicCR}(object, ...)

\method{ranef}{lsjm_classicIDM}(object, ...)

\method{ranef}{lsjm_classicSingle}(object, ...)

\method{ranef}{lsjm_covDepCR}(object, ...)

\method{ranef}{lsjm_covDepIDM}(object, ...)

\method{ranef}{lsjm_covDepSingle}(object, ...)

\method{ranef}{lsjm_interintraCR}(object, ...)

\method{ranef}{lsjm_interintraIDM}(object, ...)

\method{ranef}{lsjm_interintraSingle}(object, ...)

\method{ranef}{lsmm_classic}(object, ...)

\method{ranef}{lsmm_covDep}(object, ...)

\method{ranef}{lsmm_interintra}(object, ...)
}
\arguments{
\item{object}{Either a lsmm object or a lsjm object}

\item{...}{Further arguments}
}
\value{
A table with predicted random effects for each subjects.
}
\description{
ranef: the predicted random effects for all subjects
}
\examples{

data <- data.frame(
         ID = rep(1:5, each = 3),
         time = rep(1:3, 5),
         y = rnorm(15),
         event = rep(rbinom(5,1,0.5), each = 3),
         time_event = rep(runif(5), each = 3)
         )


m0 <- lsmm(
   formFixed = y ~ time,
   formRandom = ~ time,
   formGroup = ~ ID,
   formVar = "standard",
   timeVar = "time",
   data.long = data,
   S1 = 5,
   S2 = 5,
   nproc = 1
   )

ranef.m0 <- ranef(m0)

\donttest{

library(dplyr)

data(threeC)
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$SBP <- threeC$SBP/10
threeC <- threeC
threeC <- dplyr::group_by(threeC, ID, num.visit)
threeC <- dplyr::mutate(threeC, SBPvisit = mean(SBP))
threeC_ex1 <- threeC[!duplicated(threeC[, c("ID", "num.visit")]),
                                  c("ID", "SBPvisit", "age.visit65", "sex","dem","death",
                                    "age.first65", "age.last65","age.final65","age0_65")]

m1 <- lsmm(formFixed = SBPvisit ~ age.visit65,
               formRandom = ~ age.visit65,
               formGroup = ~ ID,
               timeVar = 'age.visit65',
               data.long = threeC_ex1,
               formVar = "cov-dependent",
               formFixedVar = ~ age.visit65+sex,
               formRandomVar = ~ age.visit65,
               correlated_re = FALSE,
               S1 = 500,
               S2 = 1000,
               nproc = 1)

l1 <- lsjm(m1,
           survival_type = 'IDM',
           formSurv_01=~1,
           formSurv_02=~sex,
           formSurv_12=~sex,
           sharedtype_01 = c("value", "variability"),
           sharedtype_02 = c("value", "slope", "variability"),
           sharedtype_12 = c("value", "variability"),
           hazardBase_01 = "Weibull",
           hazardBase_02 = "Weibull",
           hazardBase_12 = "Splines",
           delta1=~dem,
           delta2=~death,
           Time_T =~age.final65,
           Time_L =~age.last65,
           Time_R =~age.first65,
           Time_T0 =~age0_65,
           formSlopeFixed =~1,
           formSlopeRandom =~1,
           index_beta_slope = c(2),
           index_b_slope = c(2),
           nb.knots.splines = c(0,0,1),
           S1 = 1000,
           S2 = 2000,
           nproc = 10)

r1 <- ranef(m1)
r1.1 <- ranef(l1)

}

}
