% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpred.R, R/dynpred.lsjm_classicCR.R,
%   R/dynpred.lsjm_classicIDM.R, R/dynpred.lsjm_classicSingle.R,
%   R/dynpred.lsjm_covDepCR.R, R/dynpred.lsjm_covDepIDM.R,
%   R/dynpred.lsjm_covDepSingle.R, R/dynpred.lsjm_interintraCR.R,
%   R/dynpred.lsjm_interintraIDM.R, R/dynpred.lsjm_interintraSingle.R
\name{dynpred}
\alias{dynpred}
\alias{dynpred.lsjm_classicCR}
\alias{dynpred.lsjm_classicIDM}
\alias{dynpred.lsjm_classicSingle}
\alias{dynpred.lsjm_covDepCR}
\alias{dynpred.lsjm_covDepIDM}
\alias{dynpred.lsjm_covDepSingle}
\alias{dynpred.lsjm_interintraCR}
\alias{dynpred.lsjm_interintraIDM}
\alias{dynpred.lsjm_interintraSingle}
\title{dynpred: compute the dynamic predictions of an event.}
\usage{
dynpred(object, newdata, s, horizon, event, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_classicCR}(object, newdata, s, horizon, event, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_classicIDM}(object, newdata, s, horizon, event = NULL, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_classicSingle}(object, newdata, s, horizon, event = NULL, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_covDepCR}(object, newdata, s, horizon, event, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_covDepIDM}(object, newdata, s, horizon, event = NULL, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_covDepSingle}(object, newdata, s, horizon, event = NULL, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_interintraCR}(object, newdata, s, horizon, event, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_interintraIDM}(object, newdata, s, horizon, event = NULL, CI = 95, nb.draws = 1000)

\method{dynpred}{lsjm_interintraSingle}(object, newdata, s, horizon, event = NULL, CI = 95, nb.draws = 1000)
}
\arguments{
\item{object}{An \code{lsjm} object.}

\item{newdata}{A dataset containing the individuals for whom predictions are to be computed.}

\item{s}{The landmark time (a single numeric value).}

\item{horizon}{The horizon time of prediction. The function computes the probability of
experiencing the event between \code{s} and \code{horizon}. This can be a vector or a single value.}

\item{event}{An integer indicating for which event the prediction is computed. In the case of a *
competing risks model, it can be 1 or 2, whereas for other models it can only be 1.}

\item{CI}{An integer between 0 and 100 indicating the confidence level (default: \code{95} for a
95\% CI). If \code{CI = NULL}, no confidence interval is computed.}

\item{nb.draws}{An integer giving the number of Monte Carlo draws used to compute the confidence
interval (default: \code{1000}).}
}
\value{
A dataframe containing the table of predictions (\code{table.pred}).
}
\description{
This function can be used only with an \code{lsjm} object.
Individual dynamic predictions can be computed and plotted.
They are defined as the predicted probability of having event \eqn{k} between time \eqn{s} and
\eqn{s+t}
given that the subject \eqn{i} has not experienced any event before time \eqn{s},
and knowing all marker measures collected until time \eqn{s},
denoted by \eqn{Y_i(s)}, and the set of estimated parameters. The prediction is defined for
subject \eqn{i} by:
#' \deqn{
\pi_i^{0k}(s,t;\widehat{\theta}) =
P(s<T_i<s+t, \delta_i = k \mid T_i > s,
\mathcal{Y}_i(s), \widehat{\theta})
=
\frac{
  \int
  \left[
    \int_s^{s+t}
    \exp\left(
      -\sum_{c=1}^K \Lambda_{i}^{0c}(u|r_i,\widehat{\theta})
    \right)
    \lambda_{i}^{0k}(u|r_i,\widehat{\theta})
    du
  \right]
  f(\mathcal{Y}_i(s)|r_i,\widehat{\theta})
  f(r_i|\widehat{\theta})
  dr_i
}{
  \int
  \exp\left(
    -\sum_{c=1}^K \Lambda_{i}^{0c}(s|r_i,\widehat{\theta})
  \right)
  f(\mathcal{Y}_i(s)|r_i,\widehat{\theta})
  f(r_i|\widehat{\theta})
  dr_i
}
}

For illness-death and competing risks models, \eqn{K = 2}, whereas for a model with a single
event, \eqn{K = 1}.
For illness-death and single-event models, only the prediction from state (0) to state (1) can be
computed, while for competing risks, transitions (0→1) and (0→2) can both be predicted.

As in the estimation procedure, the integral over the random effects is computed by QMC
approximation and the integral over time by the Gauss-Kronrod quadrature.

The 95\\% confidence interval of predictions is obtained by the following Monte Carlo algorithm,
which can be computationally intensive. For \eqn{L} large enough and \eqn{l = 1,...,L}
(e.g. \eqn{L = 1000}):
\itemize{
\item Generate
\eqn{\widetilde{\theta}^{(l)} \sim
  \mathcal{N}(\widehat{\theta}, V(\widehat{\theta}))},
where \eqn{V(\widehat{\theta})} is the inverse of the Hessian
matrix at \eqn{\widehat{\theta}};

\item Compute
\eqn{\widetilde{\pi}^{(l)}_i(s,t;\widetilde{\theta}^{(l)})}
from the equation above;

\item Compute the 95\\% confidence interval from the 2.5th and
97.5th percentiles of the \eqn{L}-sample of
\eqn{\widetilde{\pi}^{(l)}_i(s,t;\widetilde{\theta}^{(l)})}.
}
}
\examples{

set.seed(123)
data <- data.frame(
         ID = rep(1:100, each = 3),
         time = rep(1:3, 100),
         y = rnorm(300),
         event = rep(rbinom(100,1,0.5), each = 3),
         time_event = rep(runif(100), each = 3)
         )


m0 <- lsmm(
   formFixed = y ~ time,
   formRandom = ~ time,
   formGroup = ~ ID,
   formVar = "standard",
   timeVar = "time",
   data.long = data,
   S1 = 5,
   S2 = 5,
   nproc = 1
   )

 fit <- lsjm(
      Objectlsmm = m0,
      survival_type = "Single",
      formSurv_01 = ~ 1,
      sharedtype_01 = "value",
      hazardBase_01 = "Weibull",
      delta1 = ~ event,
      Time_T = ~ time_event,
      S1 = 10,
      S2 = 10,
      nproc = 1
      )

 ind1 <- data[which(data$ID == 1),]
 dynpEvent <- dynpred(fit, ind1,  s = 2, horizon = seq(2.1,3,0.1),
                  event = 1, nb.draws = 100)


\donttest{
library(dplyr)

# Begin by running the examples from the lsjm function (see ?lsjm).
data(threeC)
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$SBP <- threeC$SBP/10


# Example: prediction of the risk to be diagnosed with dementia
# (dynpDementia) and to die (dynpDeath) between 85 and 95 years old
# given the blood pressure measurements before 85 years old for
# individual "10003", with a competing risk model
threeC_ex2 <- threeC[,c("ID", "SBP", "age.visit65", "age0_65",
                        "age.final65", "age.last65", "age.first65",
                        "dem", "death", "sex","num.visit")]

threeC_ex2$age65_CR <- NA
threeC_ex2$age65_CR[which(threeC_ex2$dem == 1)] <-
  (threeC_ex2$age.last65[which(threeC_ex2$dem == 1)] +
     threeC_ex2$age.first65[which(threeC_ex2$dem == 1)])/2
threeC_ex2$age65_CR[which(threeC_ex2$dem == 0)] <-
  threeC_ex2$age.final65[which(threeC_ex2$dem == 0)]

threeC_ex2$demCR <- threeC_ex2$dem
threeC_ex2$deathCR <- NA
threeC_ex2$deathCR[which(threeC_ex2$dem == 1)] <- 0
threeC_ex2$deathCR[which(threeC_ex2$dem == 0)] <-
  threeC_ex2$death[which(threeC_ex2$dem == 0)]

threeC_ex2 <- threeC_ex2 \%>\% group_by(ID) \%>\% dplyr::filter(age.visit65 <= age65_CR)

threeC_ex2 <- threeC_ex2[,c("ID","SBP","age.visit65","num.visit","age0_65",
                            "demCR","deathCR","age65_CR","sex")]

ind2 <- threeC_ex2[which(threeC_ex2$ID == 3),]

m2 <- lsmm(formFixed = SBP ~ age.visit65,
               formRandom = ~ age.visit65,
               formGroup = ~ ID,
               timeVar = 'age.visit65',
               data.long = threeC_ex2,
               formVar = "inter-intra",
               random_inter = TRUE,
               random_intra = TRUE,
               formGroupVisit = ~num.visit,
               correlated_re = FALSE,
               S1 = 500,
               S2 = 1000,
               nproc = 1)


l2 <- lsjm(Objectlsmm = m2,
           survival_type = 'CR',
           formSurv_01 = ~ sex,
           formSurv_02 = ~ sex,
           sharedtype_01 = c("value", "variability inter"),
           sharedtype_02 = c("value", "variability inter",
                             "variability intra"),
           hazardBase_01 = "Weibull",
           hazardBase_02 = "Weibull",
           delta1 = ~ demCR,
           delta2 = ~ deathCR,
           Time_T0 = ~ age0_65,
           Time_T = ~ age65_CR,
           nproc = 5,
           S1 = 1000,
           S2 = 2000)


dynpDementia <- dynpred(l2, ind2,  s = 2, horizon = seq(2.1,3,0.1),
                  event = 1, nb.draws = 1000)

dynpDeath <- dynpred(l2, ind2,  s = 2, horizon = seq(2.1,3,0.1),
                  event = 2, nb.draws = 1000)
}


}
