% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpred.R, R/dynpred.lsjm_classicCR.R,
%   R/dynpred.lsjm_classicIDM.R, R/dynpred.lsjm_classicSingle.R,
%   R/dynpred.lsjm_covDepCR.R, R/dynpred.lsjm_covDepIDM.R,
%   R/dynpred.lsjm_covDepSingle.R, R/dynpred.lsjm_interintraCR.R,
%   R/dynpred.lsjm_interintraIDM.R, R/dynpred.lsjm_interintraSingle.R
\name{dynpred}
\alias{dynpred}
\alias{dynpred.lsjm_classicCR}
\alias{dynpred.lsjm_classicIDM}
\alias{dynpred.lsjm_classicSingle}
\alias{dynpred.lsjm_covDepCR}
\alias{dynpred.lsjm_covDepIDM}
\alias{dynpred.lsjm_covDepSingle}
\alias{dynpred.lsjm_interintraCR}
\alias{dynpred.lsjm_interintraIDM}
\alias{dynpred.lsjm_interintraSingle}
\title{Dynamic Predictions}
\usage{
dynpred(object, newdata, s, horizon, event, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_classicCR}(object, newdata, s, horizon, event, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_classicIDM}(object, newdata, s, horizon, event = NULL, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_classicSingle}(object, newdata, s, horizon, event = NULL, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_covDepCR}(object, newdata, s, horizon, event, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_covDepIDM}(object, newdata, s, horizon, event = NULL, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_covDepSingle}(object, newdata, s, horizon, event = NULL, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_interintraCR}(object, newdata, s, horizon, event, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_interintraIDM}(object, newdata, s, horizon, event = NULL, IC = 95, nb.draws = 1000)

\method{dynpred}{lsjm_interintraSingle}(object, newdata, s, horizon, event = NULL, IC = 95, nb.draws = 1000)
}
\arguments{
\item{object}{Model object.}

\item{newdata}{Optional new data for prediction.}

\item{s}{Landmark time.}

\item{horizon}{Horizon time.}

\item{event}{Event indicator.}

\item{IC}{Confidence interval level.}

\item{nb.draws}{Number of Monte Carlo draws.}
}
\description{
Generic function for dynamic predictions.
}
