% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsjm.R
\name{lsjm}
\alias{lsjm}
\title{lsjm : Estimation of a location-scale joint model for longitudinal data with a flexible subject-specific variability and a complex survival design.}
\usage{
lsjm(
  Objectlsmm,
  survival_type = c("Single", "CR", "IDM"),
  formSurv_01,
  formSurv_02 = NULL,
  formSurv_12 = NULL,
  sharedtype_01,
  sharedtype_02 = NULL,
  sharedtype_12 = NULL,
  hazardBase_01,
  hazardBase_02 = NULL,
  hazardBase_12 = NULL,
  delta1,
  delta2 = NULL,
  Time_T,
  Time_L = NULL,
  Time_R = NULL,
  Time_T0 = NULL,
  formSlopeFixed = NULL,
  formSlopeRandom = NULL,
  index_beta_slope = NULL,
  index_b_slope = NULL,
  nb.knots.splines = c(1, 1, 1),
  nb_pointsGK = 15,
  S1 = 1000,
  S2 = 5000,
  binit = NULL,
  nproc = 1,
  clustertype = "SOCK",
  maxiter = 500,
  print.info = FALSE,
  file = NULL,
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 1e-04
)
}
\arguments{
\item{Objectlsmm}{The result of the lsmm function}

\item{survival_type}{a character defining which type of survival scheme to use either 'Single', or 'CR' or 'IDM'}

\item{formSurv_01}{one-sided formula providing on the right-side the regression variable of the risk function for transition 0-1}

\item{formSurv_02}{one-sided formula providing on the right-side the regression variable of the risk function for transition 0-2}

\item{formSurv_12}{one-sided formula providing on the right-side the regression variable of the risk function for transition 1-2}

\item{sharedtype_01}{a vector indicating the form(s) of the dependence structure. If the longitudinal model estimated is a standard mixed model, it should be included in c("value", "slope"). If it is a location-scale mixed model with a covariate and time-dependent variability one can add "current variability" and if it a model distinguishing within from between visits variabilities one can add c("inter visit variability", "intra visit variability") if there are subject-specifics.}

\item{sharedtype_02}{a vector indicating the form(s) of the dependence structure. If the longitudinal model estimated is a standard mixed model, it should be included in c("value", "slope"). If it is a location-scale mixed model with a covariate and time-dependent variability one can add "current variability" and if it a model distinguishing within from between visits variabilities one can add c("inter visit variability", "intra visit variability") if there are subject-specifics.}

\item{sharedtype_12}{a vector indicating the form(s) of the dependence structure. If the longitudinal model estimated is a standard mixed model, it should be included in c("value", "slope"). If it is a location-scale mixed model with a covariate and time-dependent variability one can add "current variability" and if it a model distinguishing within from between visits variabilities one can add c("inter visit variability", "intra visit variability") if there are subject-specifics.}

\item{delta1}{a one-sided formula with on the right-side the variable corresponding to the indicator of interest event (1 for subject experimenting the event and 0 for others).}

\item{delta2}{a one-sided formula with on the right-side the variable corresponding to the indicator $\delta_{1i}$ of second event (1 for subject experimenting the event and 0 for others).}

\item{Time_T}{a one-sided formula with on the right-side the variable corresponding to the time of event}

\item{Time_L}{In case of an IDM model: a one-side formula with a variable giving the time of last time in state (0)}

\item{Time_R}{In case of an IDM model: a one-side formula with a variable providing the first time time known in state (1)}

\item{Time_T0}{In case of delayed entry: is a a one-sided formula with on the right-side the variable corresponding to the time of entry}

\item{formSlopeFixed}{a one-sided formula corresponding to the derivative with respect to time of the function of fixed effects (if 'slope' is in sharedtype)}

\item{formSlopeRandom}{a one-sided formula corresponding to the derivative with respect to time of the function of random effects (if 'slope' is in sharedtype)}

\item{index_beta_slope}{a vector providing the index of beta parameters from the mixed model used in the slope}

\item{index_b_slope}{a vector providing the index of b parameters from the mixed model used in the slope}

\item{nb.knots.splines}{a vector providing the number of internal knots for Spline base hazard functions}

\item{nb_pointsGK}{the number of points for Gauss-Kronrod approximation : choice between 7 and 15. 15 by default.}

\item{S1}{An integer : the number of QMC draws for the first step}

\item{S2}{An integer : the number of QMC draws for the second step}

\item{binit}{optional initials parameters.}

\item{nproc}{An integer : the number of processors for parallel computing}

\item{clustertype}{one of the supported types from \code{makeCluster} function}

\item{epsd}{}
}
\description{
This function fits joint models in which
we suppose that the variance of the residual error is subject-specific and the survival submodel contains either a single event, or competing risks or an illness-death model.
Nine differents models can be estimated (see details below).
Parameters are estimated through a maximum likelihood method, using a Marquardt-Levenberg algorithm.
}
