% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmm.R
\name{lsmm}
\alias{lsmm}
\title{lsmm : Estimation of linear mixed model for longitudinal data with a flexible subject-specific variability.}
\usage{
lsmm(
  formFixed,
  formRandom,
  formGroup,
  timeVar,
  formVar = "classic",
  formFixedVar = NULL,
  formRandomVar = NULL,
  random_inter = F,
  random_intra = F,
  formGroupVisit = NULL,
  correlated_re = F,
  data.long,
  S1 = 500,
  S2 = 5000,
  nproc = 1,
  clustertype = "SOCK",
  maxiter = 100,
  print.info = FALSE,
  file = "",
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 1e-04,
  binit = NULL
)
}
\arguments{
\item{formFixed}{A formula for the fixed effects of the longitudinal submodel}

\item{formRandom}{A formula for the random effects of the longitudinal submodel}

\item{formGroup}{A formula which indicates the group variable}

\item{timeVar}{A character which indicates the time variable}

\item{formVar}{A character : type of variability either 'classic' for a classical linear mixed model or 'cov-dependent' for a subject-specific and covariate-dependent covariable variability or 'inter-intra' for distinguishing inter-visit from intra-visit variability}

\item{formFixedVar}{A formula for the fixed effects of the variance predictor if formVar == 'cov-dependent'}

\item{formRandomVar}{A formula for the random effects of the variance predictor if formVar == 'cov-dependent'}

\item{random_inter}{A logical indicating if the inter-visits variability is subject-specific when formVar = 'inter-intra'}

\item{random_intra}{A logical indicating if the intra-visit variability is subject-specific when formVar = 'inter-intra'}

\item{formGroupVisit}{A formula which indicates the visit indicator variable  when formVar = 'inter-intra'}

\item{correlated_re}{A logical indicating if the random effects of the trend are correlated to the random effects of the variability when formVar is in c('cov-dependent', 'inter-intra')}

\item{data.long}{A dataframe with the longitudinal data}

\item{S1}{An integer : the number of QMC draws for the first step}

\item{S2}{An integer : the number of QMC draws for the second step}

\item{nproc}{An integer : the number of processors for parallel computing}

\item{clustertype}{one of the supported types from \code{makeCluster} function}

\item{maxiter}{optional maximum number of iterations for the marqLevAlg iterative algorithm.}

\item{print.info}{logical indicating if the outputs of each iteration should be written}

\item{file}{optional character giving the name of the file where the outputs of each iteration should be written (if print.info=TRUE)}

\item{epsa}{optional threshold for the convergence criterion based on the parameter stability.}

\item{epsb}{optional threshold for the convergence criterion based on the objective function stability.}

\item{epsd}{optional threshold for the relative distance to maximum. This criterion has the nice interpretation of estimating the ratio of the approximation error over the statistical error, thus it can be used for stopping the iterative process whatever the problem.}

\item{binit}{optional initials parameters.}
}
\value{
A lsmm object which contains the following elements :
\describe{
\item{\code{table.res}}{The table of results : Estimation and SE}
\item{\code{result_step1}}{A marqLevAlg object with the results of the first step estimation.}
\item{\code{result_step1}}{A marqLevAlg object with the results of the second step estimation.}#'
\item{\code{info_conv_step2}}{Information about convergence (criteria and time)}
\item{\code{control}}{A list of control elements}

}
}
\description{
This function fits linear mixed effects models in which
we suppose that the variance of the residual error is subject-specific. Three differents models can be estimated (see details below)
Parameters are estimated through a maximum likelihood method, using a Marquardt-Levenberg algorithm.
}
\examples{

\donttest{

First example : a linear mixed model with subject-specific time-dependent and covariate-dependent variability

Second example : a linear mixed model with subject-specific inter-visits and intra-visits variabilities

Third example : a linear mixed model with subject-specific inter-visits variability and constant intra-visit variability

}






}
