% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmm.R
\name{lsmm}
\alias{lsmm}
\title{lsmm : Estimation of a linear mixed model for longitudinal data with a flexible subject-specific variability.}
\usage{
lsmm(
  formFixed,
  formRandom,
  formGroup,
  timeVar,
  formVar = "standard",
  formFixedVar = NULL,
  formRandomVar = NULL,
  random_inter = F,
  random_intra = F,
  formGroupVisit = NULL,
  correlated_re = F,
  data.long,
  S1 = 500,
  S2 = 5000,
  nproc = 1,
  clustertype = "SOCK",
  maxiter = 100,
  print.info = FALSE,
  file = "",
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 1e-04,
  binit = NULL
)
}
\arguments{
\item{formFixed}{A formula for the fixed effects of the longitudinal submodel}

\item{formRandom}{A formula for the random effects of the longitudinal submodel}

\item{formGroup}{A formula which indicates the group variable}

\item{timeVar}{A character which indicates the time variable}

\item{formVar}{A character : type of variability either 'standard' for a standard linear mixed model or 'cov-dependent' for a subject-specific and covariate-dependent covariable variability or 'inter-intra' for distinguishing inter-visit from intra-visit variability}

\item{formFixedVar}{A formula for the fixed effects of the variance predictor if formVar == 'cov-dependent'}

\item{formRandomVar}{A formula for the random effects of the variance predictor if formVar == 'cov-dependent'}

\item{random_inter}{A logical indicating if the inter-visits variability is subject-specific when formVar = 'inter-intra'}

\item{random_intra}{A logical indicating if the intra-visit variability is subject-specific when formVar = 'inter-intra'}

\item{formGroupVisit}{A formula which indicates the visit indicator variable  when formVar = 'inter-intra'}

\item{correlated_re}{A logical indicating if the random effects of the trend are correlated to the random effects of the variability when formVar is in c('cov-dependent', 'inter-intra')}

\item{data.long}{A dataframe with the longitudinal data}

\item{S1}{An integer : the number of QMC draws for the first step}

\item{S2}{An integer : the number of QMC draws for the second step}

\item{nproc}{An integer : the number of processors for parallel computing}

\item{clustertype}{one of the supported types from \code{makeCluster} function}

\item{maxiter}{optional maximum number of iterations for the marqLevAlg iterative algorithm.}

\item{print.info}{logical indicating if the outputs of each iteration should be written}

\item{file}{optional character giving the name of the file where the outputs of each iteration should be written (if print.info=TRUE)}

\item{epsa}{optional threshold for the convergence criterion based on the parameter stability.}

\item{epsb}{optional threshold for the convergence criterion based on the objective function stability.}

\item{epsd}{optional threshold for the relative distance to maximum. This criterion has the nice interpretation of estimating the ratio of the approximation error over the statistical error, thus it can be used for stopping the iterative process whatever the problem.}

\item{binit}{optional initials parameters.}
}
\value{
A lsmm object which contains the following elements :
\describe{
\item{\code{table.res}}{The table of results : Estimation and SE}
\item{\code{result_step1}}{A marqLevAlg object with the results of the first step estimation.}
\item{\code{result_step1}}{A marqLevAlg object with the results of the second step estimation.}#'
\item{\code{info_conv_step2}}{Information about convergence (criteria and time)}
\item{\code{control}}{A list of control elements}

}
}
\description{
This function fits linear mixed effects models in which
we can suppose that the variance of the residual error is subject-specific. Three differents models can be estimated (see details below).
Parameters are estimated through a maximum likelihood method, using a Marquardt-Levenberg algorithm.
}
\details{
The model is defined by:
\eqn{Y_{ij} = Y_{i}(t_{ij}) = \widetilde{Y}_i(t_{ij}) + \epsilon_{ij} = X_{ij}^{\top} \beta+Z_{ij}^{\top} b_{i}+\epsilon_{ij}},
where $X_{ij}$ and $Z_{ij}$ vectors of explanatory variables for subject $i$ at visit $j$, respectively associated with the fixed-effect vector \eqn{\beta} and the subject-specific random-effect vector \eqn{b_i}.
\enumerate{
\item Standard linear mixed model:
In this case, \eqn{b_i \sim \mathcal{N}(0,B)}, with \eqn{B} an unspecified matrix and the measurements error \eqn{\epsilon_{ij}} are independent Gaussian errors with variance \eqn{\sigma^2_{\epsilon}}.
\item Location-scale mixed model with time and/or covariate-dependent variability:
In this model we assume the following specification for the residual error:
\eqn{\epsilon_{ij} \sim \mathcal{N}(0,\sigma_i^2)} with \eqn{ \log(\sigma_i(t_{ij}))  = O_{ij}^{\top} \mu+M_{ij}^{\top} \tau_{i}}.
where $O_{ij}$ and $M_{ij}$ vectors of explanatory variables for subject $i$ at visit $j$, respectively associated with the fixed-effect vector \eqn{\mu} and the subject-specific random-effect vector \eqn{\tau_i}.
For the random effects we assume :
\eqn{\quad\left(\begin{array}{c}
             b_{i} \\
             \tau_i
             \end{array}\right) \sim N\left(\left(\begin{array}{c}
                                                  0 \\
                                                  0
                                                  \end{array}\right),\left(\begin{array}{cc}
                                                                           \Sigma_{b} & \Sigma_{\tau b} \\
                                                                           \Sigma_{\tau b}' & \Sigma_{\tau}
\end{array}\right)\right)}
}

As conventionally assumed in practice, the random effects \eqn{b_i} can be considered independent of the errors by setting \eqn{\Sigma_{\tau b} =0}.
\enumerate{
\item Location-scale mixed model distinguishing within and between visits variabilities:
In some studies, multiple measurements of a marker are collected at each time point.
For example, in medical research, 2 or 3 blood pressure readings are typically taken per visit,
and intra-visit variability can be informative. To capture this, we propose an LSMM that distinguishes within- and between-visit variabilities.
We introduce an additional level in longitudinal data, grouping repeated measurements by time.
For each subject \eqn{i} \eqn{(i=1,...,N)}, \eqn{Y_{ijl}} represents the \eqn{l}-th \eqn{(l=1,...,n_{ij})} measurement at visit \eqn{j} \eqn{(j=1,...,n_i)} and time \eqn{t_{ij}}.
We then define the following LSMM to decompose individual residual variance within and between visits:
\eqn{
\left\{
  \begin{array}{ll}
  Y_{ijl} =  \widetilde{Y}_i(t_{ij}) + \epsilon_{ij} + \nu_{ijl} = X_{ij}^{\top} \beta+Z_{ij}^{\top} b_{i}+\epsilon_{ij} + \nu_{ijl}, \\
  \epsilon_{ij} \sim \mathcal{N}(0,\sigma_i^2) \hspace{4mm} \text{with} \hspace{3mm} \log(\sigma_i)  = \mu_\sigma + \tau_{\sigma i},\\
  \nu_{ijl} \sim \mathcal{N}(0,\kappa_i^2) \hspace{3mm} \text{with} \hspace{3mm} \log(\kappa_i)  = \mu_\kappa + \tau_{\kappa i},\\
  \end{array}
  \right.}
}

The parameter \eqn{\mu_\sigma} and \eqn{\mu_\kappa} are the fixed intercepts for the between-visits and within-visit variances respectively.
The subject-specific random-effect \eqn{b_i} and \eqn{\tau_i = (\tau_{\sigma i},\tau_{\kappa i})^\top} are assumed to be Gaussian as presented previously.
}
\examples{

\donttest{

First example : a standard linear mixed model (constant residual variance, in time and between subjects)


Second example : a linear mixed model with subject-specific time-dependent and covariate-dependent variability

Third example : a linear mixed model with subject-specific inter-visits and intra-visits variabilities

Fourth example : a linear mixed model with subject-specific inter-visits variability and constant intra-visit variability

}






}
