% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plot.lsjm_classicCR.R,
%   R/plot.lsjm_classicIDM.R, R/plot.lsjm_classicSingle.R,
%   R/plot.lsjm_covDepCR.R, R/plot.lsjm_covDepIDM.R, R/plot.lsjm_covDepSingle.R,
%   R/plot.lsjm_interintraCR.R, R/plot.lsjm_interintraIDM.R,
%   R/plot.lsjm_interintraSingle.R, R/plot.lsmm_classic.R, R/plot.lsmm_covDep.R,
%   R/plot.lsmm_interintra.R
\name{plot.lsjm}
\alias{plot.lsjm}
\alias{plot.lsjm_classicCR}
\alias{plot.lsjm_classicIDM}
\alias{plot.lsjm_classicSingle}
\alias{plot.lsjm_covDepCR}
\alias{plot.lsjm_covDepIDM}
\alias{plot.lsjm_covDepSingle}
\alias{plot.lsjm_interintraCR}
\alias{plot.lsjm_interintraIDM}
\alias{plot.lsjm_interintraSingle}
\alias{plot.lsmm_classic}
\alias{plot.lsmm_covDep}
\alias{plot.lsmm_interintra}
\title{plot: Plot methods for \code{lsjm} or \code{lsmm} objects}
\usage{
\method{plot}{lsjm_classicCR}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_classicIDM}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  ObjectSmoothHazard = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_classicSingle}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_covDepCR}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_covDepIDM}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  ObjectSmoothHazard = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_covDepSingle}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_interintraCR}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_interintraIDM}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  ObjectSmoothHazard = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsjm_interintraSingle}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{lsmm_classic}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  ylim = NULL,
  xlim = NULL,
  ...
)

\method{plot}{lsmm_covDep}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  ylim = NULL,
  xlim = NULL,
  ...
)

\method{plot}{lsmm_interintra}(
  x,
  which = "long.fit",
  Objectpredict,
  break.times = NULL,
  ID.ind = NULL,
  ylim = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{lsmm} or a \code{lsjm} object}

\item{which}{A character indicating which plot must be display,}

\item{Objectpredict}{An object of the \code{predict} function}

\item{break.times}{A vector of breaking times to create windows if \code{which = 'long.fit'}}

\item{ID.ind}{A vector providing the id of subjects for whom we we wish to trace the individual
trajectory.}

\item{xlim}{the x limits (x1, x2) of the plot. The default value, NULL, indicates that the range
of the finite values to be plotted should be used.}

\item{ylim}{the y limits (y1, y2) of the plot. The default value, NULL, indicates that the range
of the finite values to be plotted should be used.}

\item{...}{Further arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}} or other methods.}

\item{ObjectSmoothHazard}{A SmoothHazard object (only for IDM survival model)}
}
\value{
A list containing one or more plots:
\describe{
\item{\code{long.fit}}{A plot comparing the mean marker predictions with the mean observed
measurements and their 95 confidence intervals.}
\item{\code{graph.traj.ind}}{One or more plots representing individual trajectories together
with their prediction intervals.}
\item{\code{survB}}{A plot comparing the predicted cumulative hazard function with either the
Nelson--Aalen estimator or an Illness--Death model.}
}
}
\description{
This function produces different plots (longitudinal and survival goodness-of-fit,
individual trajectory) of a fitted object of class \code{lsmm} or \code{lsjm}.
}
\details{
With \code{which="long.fit"}, this function allows to assess the fit of the longitudinal submodel
comparing the mean of marker predictions collected in some windows of times (defined by
\code{break.times} or using percentiles) to the mean of the observed measurements and its
95\\% confidence interval.

With \code{which="traj.ind"}, represents the individual trajectory with its prediction interval.

For a \code{lsjm} object only, with \code{which="survival.fit"}, the function
allows to assess the fit of the survival submodel. For each transition, the predicted cumulative
hazard function at each event time is computed given the predicted random effects. Then the mean
of the predicted cumulative hazard functions are compared with there Nelson-Aalen estimator in
the case of a single event or with competing risks. For an illness-death model, the predicted
cumulative hazard function for each transition is compared to an illness-death model estimated by
penalized likelihood accounting for interval censoring with the \code{SmoothHazard} package
}
\examples{

set.seed(123)

data <- data.frame(
         ID = rep(1:5, each = 3),
         time = rep(1:3, 5),
         y = rnorm(15),
         event = rep(rbinom(5,1,0.5), each = 3),
         time_event = rep(runif(5), each = 3)
         )


m0 <- lsmm(
   formFixed = y ~ time,
   formRandom = ~ time,
   formGroup = ~ ID,
   formVar = "standard",
   timeVar = "time",
   data.long = data,
   S1 = 5,
   S2 = 5,
   nproc = 1
   )

pred.m0 <- predict(m0, which = c("RE","Y"), data.long = data)

plot(m0,
     which = "long.fit",
     Objectpredict = pred.m0,
     break.times = c(0,1,2,3))

plot(m0,
     which = "traj.ind",
     Objectpredict = pred.m0,
     ID.ind = c(1,2))






\donttest{

library(dplyr)
data(threeC)
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$SBP <- threeC$SBP/10
threeC <- threeC \%>\% group_by(ID, num.visit) \%>\% mutate(SBPvisit = mean(SBP))
threeC$age65_CR <- NA
threeC$age65_CR[which(threeC$dem == 1)] <- (threeC$age.last65[which(threeC$dem == 1)]
                                       +threeC$age.first65[which(threeC$dem == 1)])/2
threeC$age65_CR[which(threeC$dem == 0)] <- threeC$age.final65[which(threeC$dem == 0)]
threeC$demCR <- threeC$dem
threeC$deathCR <- NA
threeC$deathCR[which(threeC$dem == 1)] <- 0
threeC$deathCR[which(threeC$dem == 0)]<-threeC$death[which(threeC$dem == 0)]
threeC <- threeC \%>\% group_by(ID) \%>\% filter(age.visit65 <= age65_CR)


threeC_ex1 <- threeC[!duplicated(threeC[, c("ID", "num.visit")]), c("ID",
     "SBPvisit", "age.visit65", "sex","age0_65", "demCR","deathCR",
      "age65_CR")]


# Estimation of a lsmm with time-dependent variability model:
m2 <- lsmm(formFixed = SBPvisit ~ age.visit65,
               formRandom = ~ age.visit65,
               formGroup = ~ ID,
               timeVar = 'age.visit65',
               data.long = threeC_ex1,
               formVar = "cov-dependent",
               formFixedVar = ~ age.visit65+sex,
               formRandomVar = ~ age.visit65,
               correlated_re = FALSE,
               S1 = 500,
               S2 = 1000,
               nproc = 1)

# Predictions:
pred.m2 <- predict(m2, which = c("RE", "Y"))

#Plot:
plot(m2, which = "long.fit", Objectpredict = pred.m2,
                           break.times = (seq(65,95,by = 2.5)-65)/10)
plot(m2, which = "traj.ind", Objectpredict = pred.m2,
                                            ID.ind = c(3,120))

# Estimation of a lsjm with time-dependent variability model and
# competing events:
l2 <- lsjm(Objectlsmm = m2,
           survival_type = 'CR',
           formSurv_01 = ~ sex,
           formSurv_02 = ~ sex,
           sharedtype_01 = c("value", "variability"),
           sharedtype_02 = c("value", "variability"),
           hazardBase_01 = "Weibull",
           hazardBase_02 = "Weibull",
           delta1 = ~ demCR,
           delta2 = ~ deathCR,
           Time_T0 = ~ age0_65,
           Time_T = ~ age65_CR,
           nproc = 5,
           S1 = 1000,
           S2 = 2000)

# Predictions:
pred.l2 <- predict(l2, which = c("RE", "Y", "Cum"))

# Plot:
plot(l2, which = "long.fit", Objectpredict = pred.l2,
                         break.times = (seq(65,95,by = 2.5)-65)/10)
plot(l2, which = "traj.ind", Objectpredict = pred.l2,
                          ID.ind = c(3,120))
plot(l2, which = "survival.fit", Objectpredict = pred.l2)
}


}
