% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict.lsjm_classicCR.R,
%   R/predict.lsjm_covDepCR.R, R/predict.lsjm_covDepIDM.R,
%   R/predict.lsjm_covDepSingle.R, R/predict.lsjm_interintraCR.R,
%   R/predict.lsjm_interintraIDM.R, R/predict.lsmm_classic.R,
%   R/predict.lsmm_covDep.R, R/predict.lsmm_interintra.R
\name{predict}
\alias{predict}
\alias{predict.lsjm_classicCR}
\alias{predict.lsjm_covDepCR}
\alias{predict.lsjm_covDepIDM}
\alias{predict.lsjm_covDepSingle}
\alias{predict.lsjm_interintraCR}
\alias{predict.lsjm_interintraIDM}
\alias{predict.lsmm_classic}
\alias{predict.lsmm_covDep}
\alias{predict.lsmm_interintra}
\title{Prediction of some quantities for each subjects}
\usage{
\method{predict}{lsjm_classicCR}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_covDepCR}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_covDepIDM}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_covDepSingle}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_interintraCR}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsjm_interintraIDM}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsmm_classic}(object, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsmm_covDep}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)

\method{predict}{lsmm_interintra}(Objectlsmm, which = "RE", Objectranef = NULL, data.long = NULL)
}
\arguments{
\item{which}{A vector of characters to indicate which predictions are computed. "RE" corresponds to the random effects, "Y" to the marker and "Cum" to the cumulative risk function(s) (only in the case of a lsjm object)}

\item{data.long}{A dataframe which contains the longitudinal data for making predictions.}

\item{object}{Either a lsmm object or a lsjm object}
}
\value{
A table for each type of prediction (RE/Y/Cum)
\describe{
\item{\code{predictRE}}{A table with predicted random effects for each subjects}
\item{\code{predictY}}{A table with predicted marker and residual variability for each subjects and at each measurement time}
\item{\code{predictCum...}}{to be completed}
}
}
\description{
This function computes different predicted quantities. For a \code{lsmm} object, it is possible to compute, for each subject, their random effects, their value of the marker and the residual variability (for each measurement time)).
For a \code{lsjm} object we could also compute cumulative hazard function for each risk transition.
}
\examples{
\donttest{#'
data(threeC)
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$SBP <- threeC$SBP/10
threeC <- threeC
threeC <- threeC \%>\% group_by(ID, num.visit) \%>\% mutate(SBPvisit = mean(SBP))
threeC_ex1 <- threeC[!duplicated(threeC[, c("ID", "num.visit")]), c("ID", "SBPvisit", "age.visit65", "sex")]

m1 <- lsmm(formFixed = SBPvisit ~ age.visit65,
               formRandom = ~ age.visit65,
               formGroup = ~ ID,
               timeVar = 'age.visit65',
               data.long = threeC_ex1,
               formVar = "cov-dependent",
               formFixedVar = ~ age.visit65+sex,
               formRandomVar = ~ age.visit65,
               correlated_re = FALSE,
               S1 = 500,
               S2 = 1000,
               nproc = 1)

pred.m1 <- predict(m1, which = c("RE","Y"), data.long = threeC_ex1)

head(pred.m1$predictRE)
head(pred.m1$predictY)

l1 <- ... *to be completed*

pred.l1 <- predict(l1, which = c("RE","Y","Cum"), data.long = threeC_ex1)


}




}
