% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.lsmm_classic.R, R/ranef.lsmm_classicCR.R,
%   R/ranef.lsmm_covDep.R, R/ranef.lsmm_interintra.R
\name{ranef}
\alias{ranef}
\alias{ranef.lsmm_classic}
\alias{ranef.lsmm_classicCR}
\alias{ranef.lsmm_covDep}
\alias{ranef.lsmm_interintra}
\title{ranef : Compute the random effects of the longitudinal submodel}
\usage{
\method{ranef}{lsmm_classic}(object, ...)

\method{ranef}{lsmm_classicCR}(object, ...)

\method{ranef}{lsmm_covDep}(object, ...)

\method{ranef}{lsmm_interintra}(object, ...)
}
\arguments{
\item{object}{A lsmm or lsjm object}
}
\description{
ranef : Compute the random effects of the longitudinal submodel

ranef : Compute the random effects of the longitudinal submodel
}
