% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.lsjm_covDepCR.R,
%   R/ranef.lsjm_covDepIDM.R, R/ranef.lsjm_interintraIDM.R,
%   R/ranef.lsmm_classic.R, R/ranef.lsmm_covDep.R, R/ranef.lsmm_interintra.R
\name{ranef.lsjm_covDepCR}
\alias{ranef.lsjm_covDepCR}
\alias{ranef.lsjm_covDepIDM}
\alias{ranef.lsjm_interintraIDM}
\alias{ranef}
\alias{ranef.lsmm_classic}
\alias{ranef.lsmm_covDep}
\alias{ranef.lsmm_interintra}
\title{ranef : Compute the random effects of the longitudinal submodel}
\usage{
\method{ranef}{lsjm_covDepCR}(object, ...)

\method{ranef}{lsjm_covDepIDM}(object, ...)

\method{ranef}{lsjm_interintraIDM}(object, ...)

\method{ranef}{lsmm_classic}(object, ...)

\method{ranef}{lsmm_covDep}(object, ...)

\method{ranef}{lsmm_interintra}(object, ...)
}
\arguments{
\item{object}{A lsmm or lsjm object}
}
\description{
ranef : Compute the random effects of the longitudinal submodel
}
