% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R, R/ranef.lsjm_classicCR.R,
%   R/ranef.lsjm_classicIDM.R, R/ranef.lsjm_classicSingle.R,
%   R/ranef.lsjm_covDepCR.R, R/ranef.lsjm_covDepIDM.R,
%   R/ranef.lsjm_covDepSingle.R, R/ranef.lsjm_interintraCR.R,
%   R/ranef.lsjm_interintraIDM.R, R/ranef.lsjm_interintraSingle.R,
%   R/ranef.lsmm_classic.R, R/ranef.lsmm_covDep.R, R/ranef.lsmm_interintra.R
\name{ranef}
\alias{ranef}
\alias{ranef.lsjm_classicCR}
\alias{ranef.lsjm_classicIDM}
\alias{ranef.lsjm_classicSingle}
\alias{ranef.lsjm_covDepCR}
\alias{ranef.lsjm_covDepIDM}
\alias{ranef.lsjm_covDepSingle}
\alias{ranef.lsjm_interintraCR}
\alias{ranef.lsjm_interintraIDM}
\alias{ranef.lsjm_interintraSingle}
\alias{ranef.lsmm_classic}
\alias{ranef.lsmm_covDep}
\alias{ranef.lsmm_interintra}
\title{Random effects predictions}
\usage{
\method{ranef}{lsjm_classicCR}(object, ...)

\method{ranef}{lsjm_classicIDM}(object, ...)

\method{ranef}{lsjm_classicSingle}(object, ...)

\method{ranef}{lsjm_covDepCR}(object, ...)

\method{ranef}{lsjm_covDepIDM}(object, ...)

\method{ranef}{lsjm_covDepSingle}(object, ...)

\method{ranef}{lsjm_interintraCR}(object, ...)

\method{ranef}{lsjm_interintraIDM}(object, ...)

\method{ranef}{lsjm_interintraSingle}(object, ...)

\method{ranef}{lsmm_classic}(object, ...)

\method{ranef}{lsmm_covDep}(object, ...)

\method{ranef}{lsmm_interintra}(object, ...)
}
\arguments{
\item{object}{Either a lsmm object or a lsjm object}

\item{...}{Further arguments passed to methods.}
}
\description{
Random effects predictions
}
