% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_marg.R, R/surv_marg.lsjm_classicCR.R,
%   R/surv_marg.lsjm_classicIDM.R, R/surv_marg.lsjm_classicSingle.R,
%   R/surv_marg.lsjm_covDepCR.R, R/surv_marg.lsjm_covDepIDM.R,
%   R/surv_marg.lsjm_covDepSingle.R, R/surv_marg.lsjm_interintraCR.R,
%   R/surv_marg.lsjm_interintraIDM.R, R/surv_marg.lsjm_interintraSingle.R
\name{survmarg}
\alias{survmarg}
\alias{survmarg.lsjm_classicCR}
\alias{survmarg.lsjm_classicIDM}
\alias{survmarg.lsjm_classicSingle}
\alias{survmarg.lsjm_covDepCR}
\alias{survmarg.lsjm_covDepIDM}
\alias{survmarg.lsjm_covDepSingle}
\alias{survmarg.lsjm_interintraCR}
\alias{survmarg.lsjm_interintraIDM}
\alias{survmarg.lsjm_interintraSingle}
\title{The population-average probability of being free of any events at a given time $t$ and covariates}
\usage{
\method{survmarg}{lsjm_classicCR}(object, individual, time)

\method{survmarg}{lsjm_classicIDM}(object, individual, time)

\method{survmarg}{lsjm_classicSingle}(object, individual, time)

\method{survmarg}{lsjm_covDepCR}(object, individual, time)

\method{survmarg}{lsjm_covDepIDM}(object, individual, time)

\method{survmarg}{lsjm_covDepSingle}(object, individual, time)

\method{survmarg}{lsjm_interintraCR}(object, individual, time)

\method{survmarg}{lsjm_interintraIDM}(object, individual, time)

\method{survmarg}{lsjm_interintraSingle}(object, individual, time)
}
\arguments{
\item{object}{A lsjm object}

\item{individual}{Subject profile for which we want to compute the probability}

\item{time}{a numeric, the time $t$ at which we want to compute the probability}
}
\value{
A numeric, the population-average probability of being free of any events at a given time $t$ and chosen covariates
}
\description{
The population-average probability of being free of any events at a given time $t$ and covariates
}
\examples{

set.seed(123)

data <- data.frame(
         ID = rep(1:100, each = 3),
         time = rep(1:3, 100),
         y = rnorm(300),
         event = rep(rbinom(100,1,0.5), each = 3),
         time_event = rep(runif(100), each = 3)
         )


m0 <- lsmm(
   formFixed = y ~ time,
   formRandom = ~ time,
   formGroup = ~ ID,
   formVar = "standard",
   timeVar = "time",
   data.long = data,
   S1 = 5,
   S2 = 5,
   nproc = 1
   )

 fit <- lsjm(
      Objectlsmm = m0,
      survival_type = "Single",
      formSurv_01 = ~ 1,
      sharedtype_01 = "value",
      hazardBase_01 = "Weibull",
      delta1 = ~ event,
      Time_T = ~ time_event,
      S1 = 100,
      S2 = 100,
      nproc = 1
      )

individual <- data[1, ]
survmarg(fit, individual, time = 2)


\donttest{

# Begining by estimating the examples from \code{lsmm} and \code{lsjm}, we then compute
# the average-probability of being alive and without dementia at age 85 for a Women

data(threeC)
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$SBP <- threeC$SBP/10
threeC <- threeC
threeC <- dplyr::group_by(threeC, ID, num.visit)
threeC <- dplyr::mutate(threeC, SBPvisit = mean(SBP))
threeC_ex1 <- threeC[!duplicated(threeC[, c("ID", "num.visit")]),
                                  c("ID", "SBPvisit", "age.visit65", "sex","dem","death",
                                    "age.first65","age.last65","age.final65","age0_65")]

m1 <- lsmm(formFixed = SBPvisit ~ age.visit65,
               formRandom = ~ age.visit65,
               formGroup = ~ ID,
               timeVar = 'age.visit65',
               data.long = threeC_ex1,
               formVar = "cov-dependent",
               formFixedVar = ~ age.visit65+sex,
               formRandomVar = ~ age.visit65,
               correlated_re = FALSE,
               S1 = 500,
               S2 = 1000,
               nproc = 1)

l1 <- lsjm(m1,
           survival_type = 'IDM',
           formSurv_01=~1,
           formSurv_02=~sex,
           formSurv_12=~sex,
           sharedtype_01 = c("value", "variability"),
           sharedtype_02 = c("value", "slope", "variability"),
           sharedtype_12 = c("value", "variability"),
           hazardBase_01 = "Weibull",
           hazardBase_02 = "Weibull",
           hazardBase_12 = "Splines",
           delta1=~dem,
           delta2=~death,
           Time_T =~age.final65,
           Time_L =~age.last65,
           Time_R =~age.first65,
           Time_T0 =~age0_65,
           formSlopeFixed =~1,
           formSlopeRandom =~1,
           index_beta_slope = c(2),
           index_b_slope = c(2),
           nb.knots.splines = c(0,0,1),
           S1 = 1000,
           S2 = 2000,
           nproc = 10)

individual <- threeC_ex1[1, c("Sex")]
survmarg(l1, individual, time = 2)


}




}
